refactor(compose): move quick profile into separate file
